RAG-With-Dockling/requirements.txt
llama-index-core 
llama-index-readers-docling 
llama-index-node-parser-docling 
llama-index-embeddings-huggingface 
llama-index-llms-huggingface-api 
llama-index-readers-file 
python-dotenv 
llama-index-llms-ollama
llama_index.llms.groq